+++ pkg/peertls/tlsopts/cert.go
@@ -1,7 +1,7 @@
 // Copyright (C) 2019 Storj Labs, Inc.
 // See LICENSE for copying information.
 
-package server
+package tlsopts
 
 const (
 	// DefaultPeerCAWhitelist includes the production Storj network CAs
