+++ pkg/server/server.go
@@ -11,6 +11,7 @@ import (
 	"google.golang.org/grpc"
 
 	"storj.io/storj/pkg/identity"
+	"storj.io/storj/pkg/peertls/tlsopts"
 )
 
 // Service represents a specific gRPC method collection to be registered
@@ -31,8 +32,8 @@ type Server struct {
 
 // New creates a Server out of an Identity, a net.Listener,
 // a UnaryServerInterceptor, and a set of services.
-func New(opts *Options, lis net.Listener, interceptor grpc.UnaryServerInterceptor, services ...Service) (*Server, error) {
-	grpcOpts, err := opts.grpcOpts()
+func New(opts *tlsopts.Options, lis net.Listener, interceptor grpc.UnaryServerInterceptor, services ...Service) (*Server, error) {
+	grpcOpts, err := opts.GRPCOpts()
 	if err != nil {
 		return nil, err
 	}
